Azure/rss-worker/requirements.txt
# Azure SDKs
azure-servicebus>=7.11.0
azure-cosmos>=4.5.0
azure-identity>=1.15.0

# HTTP and RSS
feedparser>=6.0.10
aiohttp>=3.9.0
requests>=2.31.0

# AI/ML (for embeddings)
openai>=1.0.0
numpy>=1.24.0

# Web framework (for health endpoint)
fastapi>=0.109.0
uvicorn>=0.27.0

# Data Processing
pydantic>=2.5.0
python-dateutil>=2.8.2
pytz>=2023.3

# Utilities
python-dotenv>=1.0.0

